[feat] #63 - 입금/배송 상태 Label 구현
