+++ POTI-iOS/POTI-iOS/Presentation/MyPageJoin/View/StatusRowView.swift
@@ -1,0 +1,60 @@
+//
+//  StatusRowView.swift
+//  POTI-iOS
+//
+//  Created by 이서현 on 1/19/26.
+//
+
+/// 복사 아래 밑줄!!!!!!!!!!!!!!!!!
+import UIKit
+
+import SnapKit
+import Then
+
+final class StatusRowView: BaseView {
+    
+    // MARK: - UI
+    
+    private let statusLabel = UILabel()
+    
+    // MARK: - Init
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Setup
+    
+    override func setStyle() {
+        statusLabel.do {
+            $0.font = PotiFontManager.body16sb.font
+        }
+    }
+    
+    override func setUI() {
+        addSubview(statusLabel)
+    }
+    // MARK: - Layout
+    
+    override func setLayout() {
+        
+        statusLabel.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+            $0.centerY.equalToSuperview()
+        }
+    }
+    
+    func configure(depositInfo: MyPageJoinModel.PaymentInfo) {
+        statusLabel.text = depositInfo.depositStatus.text
+        statusLabel.textColor = depositInfo.depositStatus.badgeColor
+    }
+    
+    func configure(shippingInfo: MyPageJoinModel.ShippingInfo) {
+        statusLabel.text = shippingInfo.shippingStatus.text
+        statusLabel.textColor = shippingInfo.shippingStatus.badgeColor
+    }
+}
